Oss release 27.0.5 created 2021-01-27-18-20
see CHANGELOG.txt for details

Original commit sha: 44b296e2d3b79a36358b922bd62598d624be92d6

+++ renderer/Platform/Platform_X11/src/Window_X11.cpp
@@ -185,7 +185,7 @@ namespace ramses_internal
         if (m_X11WindowData.window && m_X11WindowData.display)
         {
             XDestroyWindow(m_X11WindowData.display, m_X11WindowData.window);
-            XSync(m_X11WindowData.display, true);
+            XSync(m_X11WindowData.display, static_cast<int>(true));
 
             LOG_DEBUG(CONTEXT_RENDERER, "Window_X11::destroyLastInstance XCloseDisplay");
             XCloseDisplay(m_X11WindowData.display);
